+++ src/index.ts
@@ -1,4 +1,5 @@
 /// <reference types="@pixi/mixin-get-child-by-name" />
+
 export { glTFLoader } from "./loader/gltf-loader"
 export { glTFBinaryLoader } from "./loader/gltf-binary-loader"
 export { glTFAsset } from "./gltf/gltf-asset"