+++ src/sprite/post-processing-sprite.ts
@@ -2,7 +2,7 @@ import { RenderTexture, Renderer } from "@pixi/core"
 import { DisplayObject, IDestroyOptions } from "@pixi/display"
 import { Sprite } from "@pixi/sprite"
 import { Ticker } from "@pixi/ticker"
-import { PixiVersion } from "../pixi/pixi-version"
+import { Compatibility } from "../compatibility/compatibility"
 
 export interface PostProcessingSpriteOptions {
   /**
@@ -102,10 +102,6 @@ export class PostProcessingSprite extends Sprite {
    * @param object The object to render.
    */
   renderObject(object: DisplayObject) {
-    if (PixiVersion.isLaterThan6()) {
-      this.renderer.render(object, { renderTexture: this._renderTexture })
-    } else {
-      this.renderer.render(object, this._renderTexture)
-    }
+    Compatibility.render(this.renderer, object, this.renderTexture)
   }
 }